user card id int(10)

+++ database/migrations/2021_12_17_020126_add_card_id_to_users.php
@@ -14,7 +14,7 @@ class AddCardIdToUsers extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedTinyInteger('card_id')->after('unit_id')->default(0)->comment('卡id');
+            $table->unsignedInteger('card_id')->after('unit_id')->default(0)->comment('卡id');
         });
     }
